Se realizó la integración con tienda nube, permitiendo un MVP funcional que incluye integrar clientes, crear carrier de envio tal que aparezca en la tienda como un servicio cuando se va a comprar, webhook que responde a los precios del envío conforme la tienda haga la solicitud y un webhook que reciba las notificaciones de compra de la tienda (los webhook aun no se han probado)

+++ app/Http/Controllers/editClienteController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\access_meli;
+use App\Models\access_nube;
 use App\Models\clientes;
 use Illuminate\Http\Request;
 
@@ -20,10 +21,16 @@ class editClienteController extends Controller
         $del_client = request()->input('del_client');
         
         $cliente_eliminado = clientes::where('id','=',$del_client)->first();
-        $cliente_MELI = access_meli::where('user_id','=',$cliente_eliminado)->first();
-        $cliente_MELI->delete();
-        //$cliente_NUBE = access_nube::where('user_id','=',$cliente_eliminado);
-        //$cliente_NUBE->delete();
+        
+        if ($cliente_eliminado['id_MELI']) {        /// Elimina cliente de Mercadolibre
+            $cliente_MELI = access_meli::where('user_id','=',$cliente_eliminado['id_MELI'])->first();
+            $cliente_MELI->delete();
+        }
+        if ($cliente_eliminado['id_TN']) {        /// Elimina cliente de Tienda Nube
+            $cliente_NUBE = access_nube::where('user_id','=',$cliente_eliminado['id_TN'])->first();
+            $cliente_NUBE->delete();
+        }
+        
         $cliente_eliminado->delete();
     
         //////////////////////////////////////
